% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angioData}
\alias{angioData}
\title{List of Angiosperm species}
\format{
Dataframe
}
\usage{
data(angioData)
}
\description{
Complete list of all non-hybrid species and
associated data of Lecythidaceae, Aristolochiaceae, Martyniaceae,
Cabombaceae and Begoniaceae as retrieved from POWO database with the function
\code{powoSpecies} of this package gathered at Nov 2022.
}
