% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angioGenera}
\alias{angioGenera}
\title{List of Angiosperm genera}
\format{
Dataframe
}
\usage{
data(angioGenera)
}
\description{
Complete list of all genera and associated data of
Lecythidaceae, Aristolochiaceae, Begoniaceae, Martyniaceae, Dipterocarpaceae,
and Fagaceae as retrieved from POWO database with the function
\code{powoGenera} of this package gathered at Jan 2023.
}
