% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angiospermsData}
\alias{angiospermsData}
\title{Complete list of Angiosperms species}
\format{
Dataframe
}
\usage{
data(angiospermsData)
}
\description{
Complete list of all Angiosperms non-hybrid species
and associated data as retrieved from POWO database with the function
\code{powoSpecies} of this package gathered at 11.2022.
}
