% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{begoniaceaeData}
\alias{begoniaceaeData}
\title{Complete list of Begoniaceae's species}
\format{
Dataframe
}
\usage{
data(begoniaceaeData)
}
\description{
Complete list of the Begoniaceae's species
and associated data as retrieved from POWO database with the function
powoSpecies of this package.
}
