% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{botdivmap}
\alias{botdivmap}
\title{Global map of botanical country subdivisions}
\format{
Sf Dataframe
}
\source{
\url{https://www.tdwg.org/standards/wgsrpd/}
}
\usage{
data(botdivmap)
}
\description{
Global-scale map of botanical countries (level 3) according
to the World Geographical Scheme for Recording Plant Distributions (WGSRPD)
}
