% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{botregions}
\alias{botregions}
\title{Countries and associated classification of botanical divisions}
\format{
Table in .csv format
}
\source{
\url{https://www.tdwg.org/standards/wgsrpd/}
}
\usage{
data(botregions)
}
\description{
Countries and associated classification of botanical divisions
according to the World Geographical Scheme for Recording Plant Distributions.
}
