% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POWOcodes}
\alias{POWOcodes}
\title{Lista completa das familias de angiospermas e suas respectivas uris}
\format{
Table in .csv format
}
\usage{
data(POWOcodes)
}
\description{
Lista completa de famílias botanicas baseada na APG e extraida
pela funcao do pacote taxize \code{\link{get_pow}}
}
