% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoMap.R
\name{powoMap}
\alias{powoMap}
\title{Mapping species distribution from POWO data}
\usage{
powoMap(data = NULL,
        botctrs = NULL,
        map_div = "native_to_country",
        multigen = FALSE,
        verbose = TRUE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_map",
        format = "jpg")
}
\arguments{
\item{data}{Data frame of species' global distribution at country or
botanical country level as created with the function \code{powoSpecies}.}

\item{botctrs}{Sf data frame of botanical subdivision of the World.}

\item{map_div}{Any of the two columns of native distribution from the input
data created with \code{powoSpecies}. It has to be "native_to_country"
associated to political country, or "native_to_botanical_countries" according
to the level 3 of botanical subdivision of the
\href{https://www.tdwg.org/standards/wgsrpd/}{WGSRPD}.}

\item{multigen}{Logical, if \code{TRUE}, the function will produce multiple
global maps for every single genus within the input data.}

\item{verbose}{Logical, if \code{FALSE}, the map creation steps will not be
printed in the console in full.}

\item{vir_color}{A character vector with the name or code of any of the color
palettes from \href{https://CRAN.R-project.org/package=viridis}{Viridis} package.}

\item{bre_color}{A character vector with the name or code of any of the color
palettes from \href{https://CRAN.R-project.org/package=RColorBrewer}{RColorBrewer}
package.}

\item{leg_title}{A character vector to be displayed in the output map as a
legend. Default is to create one title called \strong{SR}, which means species
richness.}

\item{dpi}{One number in the range of 72-4000 referring to the image
resolution in the format of dots per inch in the output file. Default is to
create an output with 600 dpi.}

\item{dir}{Pathway to the computer's directory, where the map file will be
saved provided that the argument \code{save} is set up in \code{TRUE}. The
default is to create a directory named \strong{results_powoMap/} and the search
results will be saved within a subfolder named by the current date.}

\item{filename}{Name of the output file to be saved. Default is to create a
file entitled \strong{global_richness_map}.}

\item{format}{A character vector related to the file format that the global
map will be saved, which can be, e.g. a Portable Document Format (.pdf),
Tag Image File Format (.tiff), or Joint Photographic Experts Group (.jpeg).
Default is to save the output in \strong{jpg} format.}
}
\value{
Global map in .pdf format and saves the output on disk.
}
\description{
It produces global-scale maps of species richness at political
country and botanical country levels for any genus or family of flowering
plants with associated species distribution data at
\href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
library(expowo)
library(taxize)

powocodes <- cbind(family = "Lecythidaceae",
                   data.frame(taxize::get_pow("Lecythidaceae")))

mapspdist <- powoSpecies(powocodes$family, powocodes$uri,
                         hybridspp = FALSE,
                         country = NULL,
                         verbose = TRUE,
                         save = TRUE,
                         dir = "results_powoSpecies/",
                         filename = "Lecythidaceae")

# To create a map according to political countries
powoMap(data = mapspdist,
        botctrs = NULL,
        map_div = "native_to_country",
        multigen = TRUE,
        verbose = FALSE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_country_map",
        format = "jpg")

# To create a map according to botanical countries subdivision
data(botdivmap)
powoMap(data = mapspdist,
        botctrs = botdivmap,
        map_div = "native_to_botanical_countries",
        multigen = TRUE,
        verbose = FALSE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_botcountry_map",
        format = "jpg")

}

}
\seealso{
\code{\link{megaGen}}

\code{\link{toptenGen}}

\code{\link{powoSpecies}}

\code{\link{powoFam}}

\code{\link{powoGenera}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
