% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accGraph.R
\name{accGraph}
\alias{accGraph}
\title{Create graphs depicting the accumulation of species discoveries}
\usage{
accGraph(inputdf = NULL,
         verbose = TRUE,
         multi = TRUE,
         save = FALSE,
         dir = "results_accGraph",
         filename = "cumulative_discovery_",
         format = "pdf")
}
\arguments{
\item{inputdf}{A dataframe generated using \code{powoSpecies} function and
containing the genus and species column and the associated information about
the protologue of each species. The species name must be binomial, i.e.
must contain both the genus name and specific epithet.}

\item{verbose}{Logical, if \code{FALSE}, a message showing each step during
the POWO search will not be printed in the console in full.}

\item{multi}{Logical. Setting to \code{FALSE} means that your dataset contains
only one genus and the function will not create a violin plot.
The default is \code{TRUE}.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk.}

\item{dir}{Pathway to the computer's directory, where the file will be saved
provided that the argument \code{save} is set up in \code{TRUE}. The default
is to create a directory named \strong{results_accGraph} and the search results
will be saved within a subfolder named after the current date.}

\item{filename}{Name of the output file to be saved. The default is to create
a file entitled \strong{cumulative_discovery_}.}

\item{format}{A character vector related to the file format of the graph
to be saved. The default is "pdf" to save the output in Portable Document
Format (.pdf), but you can also choose "jpg" to save in Joint Photographic
Experts Group (.jpg), "tiff" to save in Tag Image File Format (.tiff) or
"png" to save in Portable Network Graphics (.png).}
}
\value{
One or a list of objects of class c("gg", "ggplot").
}
\description{
Produces graphs of accumulation of species discovery for each
genus and a flat violin graph for all data provided. It is designed to
create graphs based on the search results from \code{powoSpecies}. Multiple
graphs for any different taxonomic groups within the input data can be
produced automatically in a single run, and can generate a violin plot
comparing the distribution of data of all genera included.
}
\examples{
\dontrun{

library(expowo)

accGraph(inputdf = "output",
         verbose = TRUE,
         multi = FALSE,
         save = FALSE,
         dir = "results_accGraph",
         filename = "cumulative_discovery_Cyperaceae",
         format = "pdf")
}

}
\seealso{
\code{\link{megaGen}}

\code{\link{topGen}}

\code{\link{powoSpecies}}

\code{\link{powoFam}}

\code{\link{powoGenera}}

\code{\link{powoMap}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
