% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angioData}
\alias{angioData}
\title{Complete list of Angiosperm species}
\format{
Dataframe
}
\usage{
data(angioData)
}
\description{
Complete list of all Angiosperm non-hybrid species and
associated data as retrieved from POWO database with the function
\code{powoSpecies} of this package gathered at Nov 2022
}
