% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angioGenera}
\alias{angioGenera}
\title{List of Angiosperm genera}
\format{
\code{data.frame}
}
\usage{
data(angioGenera)
}
\description{
Complete list of all genera and associated data of
Lecythidaceae, Aristolochiaceae, Begoniaceae, Martyniaceae, Dipterocarpaceae,
and Fagaceae, as retrieved from POWO database in January 2023 using the
\code{expowo}'s function \code{powoGenera}.
}
