% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megaGen.R
\name{megaGen}
\alias{megaGen}
\title{Extracts mega-diverse genera from POWO}
\usage{
megaGen(family, uri,
        treshold = NULL,
        verbose = TRUE)
}
\arguments{
\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{uri}{one or multiple URI addresses for each family to be searched in POWO.}

\item{treshold}{defined limit of species number within a genus to be considered
megadiverse.}
}
\value{
Table in data frame format.
}
\description{
It produces a data frame listing all megadiverse genera associated to
the URI addresses of angiosperm families at \href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
The treshold to be considered a megadiverse genus should be defined by a numeric
value. \href{https://doi.org/10.2307/4135449}{Frodin (2004)} suggests 500 species
as the treshold for megadiverse genera.
}
\examples{
\dontrun{
powocodes <- taxize::get_pow(c("Fabaceae", "Lecythidaceae"))
powocodes <- data.frame(powocodes)
powocodes <- cbind(family = c("Fabaceae", "Lecythidaceae"), powocodes)

resMega <- megaGen(powocodes$family, powocodes$uri,
                   treshold = 500,
                   verbose = TRUE)
                        
write.csv(resMega, "powo_megagenera_accepted_number_spp.csv", row.names=FALSE)
}

}
\author{
Debora Zuanny & Domingos Cardoso
}
