% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POWOcodes}
\alias{POWOcodes}
\title{Complete list of angiosperm families and associated URI addresses}
\format{
Table in .csv format
}
\usage{
data(POWOcodes)
}
\description{
Complete list of the
\href{https://doi.org/10.1111/boj.12385}{APG}-based
families of flowering plants and associated URI addresses as retrieved with
the function \code{\link{get_pow}} of the package
\href{https://taxize.dev}{taxize}
}
