% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POWOcodes}
\alias{POWOcodes}
\title{Complete list of angiosperm families and associated URI addresses}
\format{
Dataframe
}
\usage{
data(POWOcodes)
}
\description{
Complete list of the APG IV-based families of flowering plants
and associated URI addresses as retrieved with the function get_pow of the
package \href{https://docs.ropensci.org/taxize/index.html}{taxize}.
}
