% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoFam.R
\name{powoFam}
\alias{powoFam}
\title{Get the species number of any plant family from POWO}
\usage{
powoFam(family, uri,
        verbose = TRUE, save = TRUE, dir, filename)
}
\arguments{
\item{family}{Either one family name or a vector of multiple families that
are present in POWO.}

\item{uri}{URI address for each family to be searched in POWO.}

\item{verbose}{Logical, if \code{FALSE}, the search results will not be
printed in the console in full.}

\item{save}{Logical, if \code{FALSE}, the search results will not be saved.}

\item{dir}{Pathway to the computer's directory, where the file will be saved
provided that the argument \code{save} is set up in \code{TRUE}. The default
is to create a directory named \strong{results_powoFam} and the searched results
will be saved within a subfolder named by the current date.}

\item{filename}{Name of the output file to be saved. Default is to create a
file entitled \strong{output}.}
}
\value{
Table in .csv format that is saved on disk.
}
\description{
It produces a CSV file listing the number of species within the
target botanical families of flowering plants available at
\href{https://powo.science.kew.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
library(expowo)
library(taxize)

fam <- c("Araceae", "Sapotaceae")
powocodes <- cbind(family = fam,
                   data.frame(taxize::get_pow(fam)))

powoFam(powocodes$family, powocodes$uri,
        verbose = TRUE,
        save = TRUE,
        dir = "results_powoFam/",
        filename = "Araceae_Sapotaceae")

## Searching for the number of species in any or all flowering plant
## families, by using the URI addresses within the POWOcodes data file.

data(POWOcodes)

powoFam(POWOcodes$family, POWOcodes$uri,
        verbose = TRUE,
        save = TRUE,
        dir = "results_powoFam/",
        filename = "all_angiosperms_species_number")
}

}
\seealso{
\code{\link{megaGen}}

\code{\link{topGen}}

\code{\link{powoGenera}}

\code{\link{powoSpecies}}

\code{\link{powoMap}}

\code{\link{POWOcodes}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
