% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoGenera.R
\name{powoGenera}
\alias{powoGenera}
\title{Extracts list of genera from POWO}
\usage{
powoGenera(family, uri)
}
\arguments{
\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{uri}{one or multiple URI addresses for each family to be searched in POWO.}
}
\value{
Table in data frame format.
}
\description{
It produces a data frame listing all genera, and associated accepted
species and geographical distribution, from URI addresses of angiosperm
families at \href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
powocodes <- taxize::get_pow(c("Fabaceae", "Lecythidaceae"))
powocodes <- data.frame(powocodes)
powocodes <- cbind(family = c("Fabaceae", "Lecythidaceae"), powocodes)

resGenera <- powoGenera(powocodes$family, powocodes$uri,
                        verbose = TRUE)
                        
write.csv(resGenera, "powo_genera_accepted_number_spp.csv", row.names=FALSE)
}

}
\seealso{
\code{\link{powoSpecies}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
