% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoGenera.R
\name{powoGenera}
\alias{powoGenera}
\title{Extracts list of genera from POWO}
\usage{
powoGenera(dir, filename, family, uri,
           genus, country,
           verbose = TRUE, save = TRUE)
}
\arguments{
\item{dir}{Pathway to computer's directory, where the file will be saved if the
param "save" is set up in \code{TRUE}. Default is to create a directory
named "results_powoGenera/".}

\item{filename}{Name of the final output file. Default is to create a file
entitled "output".}

\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{uri}{One or multiple URI addresses for each family to be searched in POWO.}

\item{genus}{Either a single genus name or a vector of multiple genera
that are present in POWO. If you do not provide any genus name, then the function
will search all accepted genera known for the family names provided.}

\item{country}{Either a single country name or a vector of multiple countries.
If you provide any country name, then the function will return only the genera
with native distribution known to any of the provided country names, as available
in POWO.}

\item{verbose}{Logical, if \code{FALSE}, the search results will not be printed
in the console in full.Defaults is TRUE.}

\item{save}{Logical, if \code{FALSE}, the search results will not be saved.
Defaults is TRUE.}
}
\value{
Table in .csv format and saves the output on disk.
}
\description{
It produces a CSV file listing all genera with associated number
of accepted species and geographical distribution, from URI addresses of angiosperm
families at \href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
powocodes <- taxize::get_pow(c("Fabaceae", "Lecythidaceae"))
powocodes <- data.frame(powocodes)
powocodes <- cbind(family = c("Fabaceae", "Lecythidaceae"), powocodes)

powoGenera(dir = "results_powoGenera/",
           filename = "Fabaceae_Lecythidaceae",
           powocodes$family,
           powocodes$uri,
           verbose = TRUE,
           save = TRUE)


powoGenera(dir = "results_powoGenera/",
           filename = "Fabaceae_Lecythidaceae",
           powocodes$family, powocodes$uri,
           genus = c("Luetzelburgia", "Bertholletia"),
           country = c("Argentina", "Brazil", "French Guiana"),
           verbose = TRUE,
           save = TRUE)

}

}
\seealso{
\code{\link{powoSpecies}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
