% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoMap.R
\name{powoMap}
\alias{powoMap}
\title{Mapping species distribution from POWO}
\usage{
powoMap(data = NULL,
        map_div = "native_to_country",
        multigen = FALSE,
        verbose = TRUE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_map",
        format = "pdf")
}
\arguments{
\item{data}{A data frame created using the function \code{powoSpecies} with
species' global distribution.}

\item{map_div}{One of two columns of native distribution from the input data
created using \code{powoSpecies} function. It has to be "native_to_country"
associated to political country, and "native_to_botanical_countries"
according to the level 3 of botanical subdivision of the World, from
\link{World Geographical Scheme for Recording Plant Distributions}
(https://www.tdwg.org/standards/wgsrpd/).}

\item{multigen}{Logical, if \code{TRUE}, the map results will include
more than one genus.}

\item{verbose}{Logical, if \code{FALSE}, the map creation steps will not be
printed in the console in full.}

\item{vir_color}{A character vector with the name or code of one of the color
palettes from \href{https://CRAN.R-project.org/package=viridis}{Viridis} package.}

\item{bre_color}{A character vector with the name or code of one of the color
palettes from \href{https://CRAN.R-project.org/package=RColorBrewer}{RColorBrewer}
package.}

\item{leg_title}{A character vector to be displayed in the output map as a
legend. Default is to create one title called \strong{SR}, which means species
richness.}

\item{dpi}{One number in the range of 72-4000 referring to the image
resolution in the format of dots per inch in the output file. Default is to
create an output with 600 dpi.}

\item{dir}{Pathway to the computer's directory, where the map file will be
saved provided that the argument \code{save} is set up in \code{TRUE}. The
default is to create a directory named \strong{results_powoMap/} and the search
results will be saved within a subfolder named by the current date.}

\item{filename}{Name of the output file to be saved. Default is to create a
file entitled \strong{global_richness_map}.}

\item{format}{A character vector related to the file format that the global
map will be saved, which can be, e.g., a Portable Document Format (.pdf),
Tag Image File Format (.tiff), or Joint Photographic Experts Group (.jpeg).
Default is to save the output in \strong{jpg} format.}
}
\value{
Global map in .pdf format and saves the output on disk.
}
\description{
It produces a global-scale map associated with all accepted
species and their geographical distribution from any target genus or family
of flowering plants at
\href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
library(expowo)
library(taxize)

powocodes <- cbind(family = "Lecythidaceae",
                   data.frame(taxize::get_pow("Lecythidaceae")))

mapspdist <- powoSpecies(powocodes$family, powocodes$uri,
                         hybridspp = FALSE,
                         country = NULL,
                         verbose = TRUE,
                         save = TRUE,
                         dir = "results_powoSpecies/",
                         filename = "Lecythidaceae")

# To create a map according to political countries
powoMap(data = mapspdist,
        botctrs = NULL,
        map_div = "native_to_country",
        multigen = TRUE,
        verbose = FALSE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_country_map",
        format = "jpg")

# To create a map according to botanical countries subdivision
data(botdivmap)
powoMap(data = mapspdist,
        botctrs = botdivmap,
        map_div = "native_to_botanical_countries",
        multigen = TRUE,
        verbose = FALSE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_botcountry_map",
        format = "jpg")

}

}
\seealso{
\code{\link{megaGen}}

\code{\link{toptenGen}}

\code{\link{powoSpecies}}

\code{\link{powoFam}}

\code{\link{powoGenera}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
