% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoSpDist.R
\name{powoSpDist}
\alias{powoSpDist}
\title{Get list of species distribution from POWO}
\usage{
powoSpDist(family, uri, species = NULL,
            verbose = TRUE, save = TRUE, dir, filename)
}
\arguments{
\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{uri}{URI address for each family associated to the target genus to be
searched in POWO.}

\item{species}{Either one non hybrid species name or a vector of multiple
species that are present in POWO. If any species name is not provided, then
the function will search any species from all accepted genera known for the
target family.}

\item{verbose}{Logical, if \code{FALSE}, the search results will not be
printed in the console in full.}

\item{save}{Logical, if \code{FALSE}, the search results will not be saved
on disk.}

\item{dir}{Pathway to the computer's directory, where the file will be saved
provided that the argument \code{save} is set up in \code{TRUE}. The default
is to create a directory named \strong{results_powoSpDist} and the search
results will be saved within a subfolder named by the current date.}

\item{filename}{Name of the output file to be saved. Default is to create a
file entitled \strong{output}.}
}
\value{
Table in .csv format and saves the output on disk.
}
\description{
It produces a CSV file listing the geographical distribution
of all target species of flowering plants available at
\href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
library(expowo)
library(taxize)

fam <- c("Araceae", "Lecythidaceae")
powocodes <- cbind(family = fam,
                   data.frame(taxize::get_pow(fam)))

powoSpDist(powocodes$family, powocodes$uri,
           species = c("Anthurium longicaudatum", "Lecythis pisonis"),
           verbose = TRUE,
           save = TRUE,
           dir = "results_powoSpDist/",
           filename = "Araceae_Lecythidaceae")

## Searching for all species and associated global distribution
## in any or all flowering plant family, by using the URI addresses
## within the POWOcodes data file.

data(POWOcodes)

powoSpDist(POWOcodes$family, POWOcodes$uri,
           verbose = TRUE,
           save = TRUE,
           dir = "results_powoSpDist/",
           filename = "all_angiosperm_species_dist")
}

}
\seealso{
\code{\link{POWOcodes}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
