% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoSpecies.R
\name{powoSpecies}
\alias{powoSpecies}
\title{Extracts list of species from POWO}
\usage{
powoSpecies(family, genus, uri)
}
\arguments{
\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{genus}{Either a single genus name or a vector of multiple genera
that are present in POWO.}

\item{uri}{one or multiple URI addresses for each genus to be searched in POWO.}
}
\value{
Table in data frame format
}
\description{
It produces a data frame listing all accepted species and associated
geographical distribution, from URI addresses of genera of angiosperm
families at \href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
powocodes <- taxize::get_pow(c("Fabaceae", "Lecythidaceae"))
powocodes <- data.frame(powocodes)
powocodes <- cbind(family = c("Fabaceae", "Lecythidaceae"), powocodes)

resGenera <- powoGenera(powocodes$family, powocodes$uri,
                        verbose = TRUE)

resSpecies <- powoSpecies(resGenera$family, resGenera$genus, resGenera$uri,
                          hybridspp = FALSE,
                          verbose = TRUE)

write.csv(resSpecies, "powo_genera_list_accepted_spp.csv", row.names=FALSE)
}

}
\seealso{
\code{\link{powoGenera}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
