% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powoSpecies.R
\name{powoSpecies}
\alias{powoSpecies}
\title{Get list of species from POWO}
\usage{
powoSpecies(family, uri, genus = NULL, hybridspp = FALSE, country = NULL,
            verbose = TRUE, save = TRUE, dir, filename)
}
\arguments{
\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{uri}{URI address for each family associated to the target genus to be
searched in POWO.}

\item{genus}{Either one genus name or a vector of multiple genera
that are present in POWO. If any genus name is not provided, then the
function will search any species from all accepted genera known for the
target family.}

\item{hybridspp}{Logical, if \code{TRUE}, the search results will include
hybrid species.}

\item{country}{Either one country name or a vector of multiple countries.
If country names are provided, then the function will return only the species
that are native to such countries, according to POWO.}

\item{verbose}{Logical, if \code{FALSE}, the search results will not be
printed in the console in full.}

\item{save}{Logical, if \code{FALSE}, the search results will not be saved
on disk.}

\item{dir}{Pathway to the computer's directory, where the file will be saved
provided that the argument \code{save} is set up in \code{TRUE}. The default
is to create a directory named \strong{results_powoSpecies} and the search
results will be saved within a subfolder named by the current date.}

\item{filename}{Name of the output file to be saved. Default is to create a
file entitled \strong{output}.}
}
\value{
Table in .csv format and saves the output on disk.
}
\description{
It produces a CSV file listing all accepted species and
associated geographical distribution from any target genus or family of
flowering plants at
\href{https://powo.science.kew.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
library(expowo)
library(taxize)

fam <- c("Araceae", "Lecythidaceae")
powocodes <- cbind(family = fam,
                   data.frame(taxize::get_pow(fam)))

powoSpecies(powocodes$family, powocodes$uri,
            hybridspp = FALSE,
            country = c("Argentina", "Brazil", "French Guiana"),
            verbose = TRUE,
            save = TRUE,
            dir = "results_powoSpecies/",
            filename = "Araceae_Lecythidaceae")

## Searching for all species and associated global distribution
## in any or all flowering plant genera, by using the URI addresses
## within the POWOcodes data file

data(POWOcodes)

powoSpecies(POWOcodes$family, POWOcodes$uri,
            hybridspp = TRUE,
            verbose = TRUE,
            save = TRUE,
            dir = "results_powoSpecies/",
            filename = "all_angiosperm_species")
}

}
\seealso{
\code{\link{megaGen}}

\code{\link{topGen}}

\code{\link{powoFam}}

\code{\link{powoGenera}}

\code{\link{POWOcodes}}
}
\author{
Debora Zuanny & Domingos Cardoso
}
