% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptenGen.R
\name{toptenGen}
\alias{toptenGen}
\title{Extracts top ten genera with highest species diversity}
\usage{
toptenGen(family, uri,
          verbose = TRUE)
}
\arguments{
\item{family}{Either a single family name or a vector of multiple families
that are present in POWO.}

\item{uri}{one or multiple URI addresses for each family to be searched in POWO.}

\item{verbose}{Logical, if \code{FALSE}, the search results will not be printed
in the console in full.}
}
\value{
Table in data frame format.
}
\description{
It produces a data frame listing the top ten most diverse genera
of any target botanical family associated to the URI addresses of angiosperm
families at \href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}.
}
\examples{
\dontrun{
powocodes <- taxize::get_pow(c("Araceae", "Lecythidaceae"))
powocodes <- data.frame(powocodes)
powocodes <- cbind(family = c("Araceae", "Lecythidaceae"), powocodes)

resTopten <- toptenGen(powocodes$family, powocodes$uri,
                       verbose = TRUE)

write.csv(resTopten, "powo_toptengen_accepted_number_spp.csv", row.names=FALSE)
}

}
\author{
Debora Zuanny & Domingos Cardoso
}
